requirements_production.txt
# AccessAtlas Production Training Pipeline Requirements

# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Web scraping for tiles
requests>=2.31.0

# Optional: Advanced features
# Uncomment to enable

# Experiment tracking
# wandb>=0.15.0

# Model export
# onnx>=1.14.0
# onnxruntime>=1.15.0
# coremltools>=7.0  # For iOS export

# Additional augmentation
# albumentations>=1.3.0

# Distributed training
# accelerate>=0.20.0

# Web automation for tile generation
# selenium>=4.10.0
# webdriver-manager>=3.8.0

# Plotting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
